Rename logging.py to logger.py to avoid naming conflict in jsonlogger when in debug mode
