+++ sshmitm/server/__init__.py
@@ -27,7 +27,7 @@ from sshmitm.forwarders.tunnel import (
 )
 from sshmitm.interfaces.server import BaseServerInterface, ServerInterface
 from sshmitm.interfaces.sftp import BaseSFTPServerInterface, SFTPProxyServerInterface
-from sshmitm.logging import Colors
+from sshmitm.logger import Colors
 from sshmitm.multisocket import create_server_sock
 from sshmitm.session import Session
